Add tutorial_scheduling_layer_norm.py
